added folders

+++ server/database/DBManager.js
@@ -23,9 +23,10 @@ class DBManager {
 
     deleteFiles(dir, files) {
         if (files.length)
-            for (const file of files) {
-                unlink(path.join(dir, file), (err) => { });
-            }
+            for (const file of files)
+                if (file == '.gitkeep')
+                    unlink(path.join(dir, file), (err) => { });
+
     };
 
     async deleteAllFiles() {
